requirements.txt
apscheduler==3.11.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
deprecated==1.2.18
flask==3.1.1
flask-limiter==3.12
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
limits==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.4
rich==13.9.4
typing-extensions==4.14.0
tzlocal==5.3.1
urllib3==2.5.0
werkzeug==3.1.3
wrapt==1.17.2
